pyproject.toml
[project]
name = "slash-dice-bot"
version = "0.1.0"
dependencies = [
    "aiohappyeyeballs==2.4.3",
    "aiohttp==3.10.10",
    "aiosignal==1.3.1",
    "annotated-types==0.7.0",
    "anyio==4.6.2.post1",
    "attrs==24.2.0",
    "click==8.1.7",
    "fastapi==0.115.2",
    "frozenlist==1.4.1",
    "greenlet==3.1.1",
    "h11==0.14.0",
    "httptools==0.6.4",
    "idna==3.10",
    "multidict==6.1.0",
    "propcache==0.2.0",
    "psycopg2-binary==2.9.10",
    "py-cord[speed]==2.6.1",
    "pydantic-core==2.23.4",
    "pydantic==2.9.2",
    "python-dotenv==1.0.1",
    "pyyaml==6.0.2",
    "sniffio==1.3.1",
    "sqlalchemy==2.0.36",
    "starlette==0.40.0",
    "table2ascii==1.1.3",
    "typing-extensions==4.12.2",
    "uvicorn==0.32.0",
    "uvloop==0.21.0",
    "watchfiles==0.24.0",
    "wcwidth==0.2.13",
    "websockets==13.1",
    "yarl==1.15.5",
]
